Add files via upload
Short program to ask the user to input the number of random colors they would like printed.  
Print out the number random colors from a list of at least 20 colors.  
Each color should printed with a number and period in front as shown below.  
Colors may be repeated when printed.