Add files via upload
Short program using two tactile push buttons and an LED create the following system with RPi hardware. 

Button One:  Turns on and off the LED.   The first push turns on the LED and the second push turns off the LED.  

Button Two:   This button changes the blink cycle rate of the LED.  It cycles through one of five settings when the LED is on.  
Initially the LED is on all the time or has a cycle rate of 0 Hz.  As the button is pushed the cycle rate is changed to the setting as shown below.   
The setting should cycle back through the list as button is pushed.