Add files via upload
 Python function median(listName) that takes a list of numbers as an argument and prints the median value.  Function should work for any list of numbers.

If there is an odd number of data values then the median will be the value in the middle. If there is an even number of data values the median is the mean of the two data values in the middle. For the data set 1, 1, 2, 5, 6, 6, 9 the median is 5. For the data set 1, 1, 2, 6, 6, 9 the median is 4.